+++ src/test/java/ch/ethz/matsim/av/dynamics/AVPickupDropoffTest.java
@@ -207,10 +207,8 @@ public class AVPickupDropoffTest {
 	static private AVConfigGroup createConfig() {
 		AVConfigGroup config = new AVConfigGroup();
 
-		AVScoringParameterSet scoringParams = new AVScoringParameterSet();
-		scoringParams.setSubpopulation(null);
+		AVScoringParameterSet scoringParams = config.getScoringParameters(null);
 		scoringParams.setMarginalUtilityOfWaitingTime(-0.84);
-		config.addScoringParameters(scoringParams);
 
 		OperatorConfig operatorConfig = new OperatorConfig();
 		operatorConfig.getDispatcherConfig().setType(MultiODHeuristic.TYPE);