+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+- Do not require constant link waiting time attribute to be set for dynamic waiting time estimation
 - Add 'null' subpopulation scoring parameters by default
 - Automatically pass vehicle type to generators